Builds and Runs

+++ src/com/zandgall/plorpland/entity/octoplorp/ThrownSword.java
@@ -7,6 +7,8 @@
 
 package	com.zandgall.plorpland.entity.octoplorp;
 
+import org.joml.Matrix4f;
+
 import com.zandgall.plorpland.Main;
 import com.zandgall.plorpland.entity.Entity;
 import com.zandgall.plorpland.entity.PlantedSword;
@@ -50,7 +52,7 @@ public class ThrownSword extends Entity {
 	public void render() {
 		Shader.Image.use();
 		Shader.Image.setTexture(image.getTexture());
-		Shader.Image.setModel(new Matrix4f().translate(-1, -0.5, GLHelper.LAYER_1_DEPTH).rotateZ(rotation).translate(getX(), getY(), 0).scale(2, 1, 0));
+		Shader.Image.setModel(new Matrix4f().translate(-1, -0.5f, GLHelper.LAYER_1_DEPTH).rotateZ((float)rotation).translate((float)getX(), (float)getY(), 0).scale(2, 1, 0));
 		plaque.draw(target.x-1, target.y-1.8, 2, 2, GLHelper.LAYER_1_DEPTH);
 	}
